feat: integrate Topaz Design System components; replace existing UI elements with TopazDialog, TopazBottomNavigationMenu, and TopazScaffold
