--- app/src/main/java/com/example/droidchat/ui/components/dialog/AppDialog.kt
@@ -1,60 +1,0 @@
-package com.example.droidchat.ui.components.dialog
-
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.droidchat.ui.theme.DroidChatTheme
-
-@Composable
-internal fun AppDialog(
-    title: String? = null,
-    message: String,
-    confirmButtonText: String,
-    onEventDismiss: () -> Unit,
-    onEventConfirm: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onEventDismiss,
-        title = {
-            if (title != null)
-                Text(
-                    text = title,
-                    color = MaterialTheme.colorScheme.error,
-                    style = MaterialTheme.typography.titleMedium
-                )
-        },
-        text = {
-            Text(
-                text = message,
-                color = MaterialTheme.colorScheme.onError,
-                style = MaterialTheme.typography.bodyMedium
-            )
-        },
-        confirmButton = {
-            TextButton(
-                onClick = onEventConfirm,
-                content = { Text(text = confirmButtonText) }
-            )
-        },
-        containerColor = MaterialTheme.colorScheme.outlineVariant,
-        titleContentColor = MaterialTheme.colorScheme.onSecondary,
-        textContentColor = MaterialTheme.colorScheme.onSecondary
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun AlertDialogErrorApiPreview() {
-    DroidChatTheme {
-        AppDialog(
-            title = "Erro de validação de formulário",
-            message = "Erro de validação de formulário, confira os dados e tente novamente",
-            confirmButtonText = "Ok",
-            onEventDismiss = { },
-            onEventConfirm = { },
-        )
-    }
-}
